moved duplicate code into a component
